Settings page established. Moved 'bio' and 'theme' update to settings. Need to finish upload image
