+++ src/app/components/posts/post/post.component.html
@@ -67,14 +67,14 @@
     <mat-card-actions class="flex justify-between">
 
       <div class="flex gap-4">
-        <mat-icon class="icon" matTooltip="Like" matTooltipClass="tooltip" matTooltipPosition="above"
-                  (click)="likePost(post.id)">bookmark_border</mat-icon>
-        <mat-icon class="icon" matTooltip="Like" matTooltipClass="tooltip" matTooltipPosition="above"
+        <mat-icon class="icon" matTooltip="Bookmark" matTooltipClass="tooltip" matTooltipPosition="above"
           (click)="likePost(post.id)">thumb_up</mat-icon>
         <mat-icon class="icon" matTooltip="Dislike" matTooltipClass="tooltip" matTooltipPosition="above"
           (click)="dislikePost(post.id)">thumb_down</mat-icon>
       </div>
       <div class="flex gap-4">
+        <mat-icon class="icon" matTooltip="Like" matTooltipClass="tooltip" matTooltipPosition="above"
+                  (click)="likePost(post.id)">bookmark_border</mat-icon>
         <mat-icon class="icon" matTooltip="Share" matTooltipClass="tooltip" matTooltipPosition="above"
           (click)="sharePost()">share</mat-icon>
         <mat-icon class="icon" matTooltip="Comment" matTooltipClass="tooltip" matTooltipPosition="above"
