+++ src/app/pages/profile/profile.component.ts
@@ -66,77 +66,10 @@ export class ProfileComponent {
     })
   }
 
-  scrollTop() {
-    //this.document.documentElement.scrollTop = 0;
-  }
-
   // boolean toggle for modifying biography
   modifyProfileBio() {
     this.modifyBio = !this.modifyBio;
     console.log("toggle modifying bio to: " + this.modifyBio)
-  }
-
-  // submits form for biography
-  submitForm(): void {
-    if (!this.changeBioForm.valid) {
-      console.log("bio form not set")
-    }
-    // create a biography payload to send to the back end
-    const payload: BioPayload = {
-      profileId: this.profile.profileId,
-      bio: this.changeBioForm.controls.bio.value!,
-      themeName: ""
-    }
-    // setting bio for local test
-    this.profile.bio = payload.bio;
-
-    console.log("Session id: " + this.tokenService.getUser().id )
-
-    // send new bio to backend
-    this.profileService.updateUserBio(this.tokenService.getUser().id, payload).subscribe( {
-
-      next: (resp: any) => {
-        this.profile = resp;
-        this.theme = this.profile.themeName;
-        console.log(this.profile);
-      },
-      error: (err) => {
-        console.error("Issue with retrieving profile details.");
-        console.log("Error retrieving user with id: "  + this.user_id + " : " + err);
-      }
-    })
-
-    // leave modify after accepting
-    this.modifyProfileBio();
-
-    console.log("New bio is: " + payload.bio);
-  }
-
-  // css theme selector
-  selectTheme(choice: string) {
-    this.theme = choice;
-
-    // create a theme payload to send to the back end
-    const payload: BioPayload = {
-      profileId: this.profile.profileId,
-      bio: this.changeBioForm.controls.bio.value!,
-      themeName: choice
-    }
-
-    // send new bio to backend
-    this.profileService.updateTheme(this.tokenService.getUser().id, payload).subscribe( {
-
-      next: (resp: any) => {
-
-        console.log("Theme has successfully been set");
-        this.theme = payload.themeName;
-      },
-      error: (err) => {
-        console.error("Issue with retrieving profile details.");
-        console.log("Error retrieving user with id: "  + this.user_id + " : " + err);
-      }
-    })
-
   }
 
 
@@ -152,7 +85,6 @@ export class ProfileComponent {
     }
     this.setImageAndVideoFlags();
   }
-
 
   //
   setImageAndVideoFlags() {
