+++ src/app/pages/profile/settings/settings.component.html
@@ -1,7 +1,52 @@
 
 <div class="flex flex-col items-center mb-40 bg-gradient-to-r from-purple-500 to-blue-500 min-h-screen">
 
-  <p>Settings Page</p>
+  <button (click)="backClicked()" class="m-2 bg-amber-400 top-0 pt-10 pl-10 px-5 py-3 shadow rounded-md ">Go Back</button>
+
+    <div class="flex space-x-16">
+
+      <div class="bg-white flex flex-col gap-7 p-10 shadow-2xl rounded-lg  mt-20 w-1/3">
+        <p>Change Profile Image</p>
+      </div>
+
+      <div class="bg-white flex flex-col gap-7 p-10 shadow-2xl rounded-lg  mt-20  w-1/3">
+        <p>Change your `About Me`</p>
+        <p>{{profile.bio}}</p>
+
+        <!-- Form for submitting bio, only toggled when modify is set to true -->
+        <mat-card-header class="pb-5">
+          <mat-card class="w-full">
+            <form [formGroup]="changeBioForm" (submit)="submitForm( )">
+              <mat-form-field class="form-field flex">
+                <mat-label> {{profile.bio}} </mat-label>
+                <input matInput formControlName="bio">
+              </mat-form-field>
+              <button mat-raised-button color="accent" type="submit" class="w-full">Submit bio</button>
+            </form>
+          </mat-card>
+        </mat-card-header>
+      </div>
+
+      <div class="bg-white flex flex-col gap-7 p-10 shadow-2xl rounded-lg  mt-20  w-1/3">
+        <p>Change your theme</p>
+        <p>{{profile.themeName}}</p>
+
+        <div >
+          <!-- Button for user to start/stop modifying bio -->
+            <mat-form-field >
+              <mat-label>Select a theme</mat-label>
+              <mat-select>
+                <mat-option (click)="selectTheme('theme-one')" value="theme-one">Theme One</mat-option>
+                <mat-option (click)="selectTheme('theme-two')" value="theme-two">Theme Two</mat-option>
+                <mat-option (click)="selectTheme('theme-three')" value="theme-three">Theme Three</mat-option>
+                <mat-option (click)="selectTheme('theme-four')" value="theme-four">Theme Four</mat-option>
+                <mat-option (click)="selectTheme('theme-five') "value="theme-five">Theme Five</mat-option>
+              </mat-select>
+            </mat-form-field>
+          </div>
+      </div>
+
+    </div>
 
 </div>
